+++ libpqtypes/src/pqt_config.h
@@ -25,9 +25,6 @@
 /* Define to 1 if you have the <limits.h> header file. */
 #define HAVE_LIMITS_H 1
 
-/* Define to 1 if you have the <linux/limits.h> header file. */
-#define HAVE_LINUX_LIMITS_H 1
-
 /* Define to 1 if you have the `localtime_r' function. */
 #define HAVE_LOCALTIME_R 1